+++ src/types.rs
@@ -1040,6 +1040,7 @@ url = "https://github.com/org/base.git"
             )]),
             items: indexmap::IndexMap::new(),
             config_entries: std::collections::BTreeMap::new(),
+            dependency_model_aliases: indexmap::IndexMap::new(),
         };
         let serialized = toml::to_string_pretty(&lock).unwrap();
         assert!(serialized.contains("subpath = \"plugins/foo\""));
@@ -1219,6 +1220,7 @@ subpath = "plugins\\foo"
             )]),
             items: indexmap::IndexMap::new(),
             config_entries: std::collections::BTreeMap::new(),
+            dependency_model_aliases: indexmap::IndexMap::new(),
         };
 
         crate::lock::write(dir.path(), &lock).unwrap();